requirements.txt
pkg-resources==0.0.0
scikit-surprise==1.0.6
pandas==0.24.2